+++ ClaudeUsage/AppDelegate.swift
@@ -11,6 +11,7 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     private var hoverMonitor: Any?
 
     func applicationDidFinishLaunching(_ notification: Notification) {
+        ProcessInfo.processInfo.disableAutomaticTermination("Menu bar app must stay running")
         startBinding()
     }